switched completely to pydantic. Added hack to satisfy pedantic despite configuration not following pedantic closely. Configuration structure needs to be reviewed to improve code readability in that regard for the future.
